scraper/requirements.txt
supabase==2.3.1
selenium==4.16.0
webdriver-manager==4.0.1
python-dotenv==1.0.0
loguru==0.7.2
